# docs/swagger.yaml
definitions:
  handlers.AuthRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.Item:
    properties:
      id:
        type: integer
      image:
        description: path หรือ url
        type: string
      name:
        type: string
      price:
        type: integer
    type: object
  models.ItemDTO:
    properties:
      id:
        type: integer
      name:
        type: string
      price:
        type: integer
    type: object
info:
  contact: {}
  title: Go JWT API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Register
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Register
      tags:
      - Auth
  /items:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: price_asc | price_desc | name_asc | name_desc
        in: query
        name: sort
        type: string
      - description: Search keyword
        in: query
        name: q
        type: string
      - description: Cursor ID
        in: query
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get items (pagination, search, sort, cursor)
      tags:
      - Item
  /items/{id}/upload:
    post:
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item image
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload item image
      tags:
      - items
  /items/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Item'
      security:
      - BearerAuth: []
      summary: Create item
      tags:
      - Item
  /items/update/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ItemDTO'
            type: array
      security:
      - BearerAuth: []
      summary: Update item
      tags:
      - Item
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"